findInverse {KneeArrower}	R Documentation

Finds value of x whose y value is y0

Description

Finds value of x whose y value is y0

Usage


findInverse(x, y, y0)

Arguments

x
x coordinates

y
y coordinates

y0
value for y for which we want corresponding x value

Value

value of the x coordinate whose y value is y0. If such x isn't in the domain of the function, then this function will return NA.

[Package KneeArrower version 0.1.0 Index]